Add types, move to different files
