+++ week-1-ts-ecosystem/w01e01-typescript-basics/src/helpers.ts
@@ -1,0 +1,13 @@
+import {User} from "./types";
+
+function hasAddress(user: User): boolean {
+  return Boolean(user.address)
+}
+
+function hasGivenAge(requiredAge: number): (user: User) => boolean {
+  return (user) => user.age >= requiredAge
+}
+
+const isAdult = hasGivenAge(18)
+
+export {hasAddress, hasGivenAge, isAdult}
