+++ week-1-ts-ecosystem/w01e01-typescript-basics/src/main.ts
@@ -5,26 +5,8 @@
  * całość programu musi jednak działać tak jak do tej pory !
  *
  * */
-
-const user: any = {
-  name: 'Andy',
-  age: 30,
-  email: 'andy@mail-me-tommorow.com',
-  address: {
-    street: 'Strange Alley',
-    no: 23,
-  },
-}
-
-function hasAddress(user: any): any {
-  return Boolean(user.address)
-}
-
-function hasGivenAge(requiredAge: any): any {
-  return (user: any): any => user.age >= requiredAge
-}
-
-const isAdult = hasGivenAge(18)
+import {user} from "./model";
+import {hasAddress, isAdult} from "./helpers";
 
 console.log(`User ${user.name} is ${isAdult(user) ? 'adult' : 'minor'}`)
 console.log(`and has${hasAddress(user) ? '' : ' no'} address`)
