+++ week-1-ts-ecosystem/w01e01-typescript-basics/src/model.ts
@@ -1,0 +1,13 @@
+import {User} from "./types";
+
+const user: User = {
+  name: 'Andy',
+  age: 30,
+  email: 'andy@mail-me-tommorow.com',
+  address: {
+    street: 'Strange Alley',
+    no: 23,
+  },
+}
+
+export {user}
