+++ week-1-ts-ecosystem/w01e01-typescript-basics/src/types.ts
@@ -1,0 +1,13 @@
+interface Address {
+  street: string;
+  no: number;
+}
+
+interface User {
+  name: string;
+  age: number;
+  email: string;
+  address?: Address;
+}
+
+export {Address, User}
